refactor(subagent): split agent-manager.ts into agent-store and agent-panel
